refactor(whitePaper): remove whitePaper
removed from package, used only in documentation

BREAKING CHANGE:
Text: was built using whitePaper, the modifier values were renamed to Camelcase:
inline-block -> inlineBlock
link-minor -> linkMinor

Theme: removed unused modifier _gap

issue - #403
