--- .stylelintignore
@@ -1,1 +1,0 @@
-src/utils/whitepaper/whitepaper.css
