+++ src/components/Button/__stories__/examples/ButtonExampleName/ButtonExampleName.tsx
@@ -5,7 +5,7 @@ import { IconArrowRight } from '../../../../../icons/IconArrowRight/IconArrowRig
 import { IconEdit } from '../../../../../icons/IconEdit/IconEdit';
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cnDocsExample } from '../../../../../uiKit/components/DocsExample/DocsExample';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Button } from '../../../Button';
 
 export function ButtonExampleNameSemantics() {
