+++ src/components/Button/__stories__/examples/ButtonExampleView/ButtonExampleView.tsx
@@ -10,7 +10,7 @@ import { IconFunnel } from '../../../../../icons/IconFunnel/IconFunnel';
 import { IconSettings } from '../../../../../icons/IconSettings/IconSettings';
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cnDocsExample } from '../../../../../uiKit/components/DocsExample/DocsExample';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Button } from '../../../Button';
 
 export function ButtonExampleViewPrimary() {
