+++ src/components/ChoiceGroup/__stories__/examples/ChoiceGroupExampleText/ChoiceGroupExampleText.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cnDocsExample } from '../../../../../uiKit/components/DocsExample/DocsExample';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { ChoiceGroup } from '../../../ChoiceGroup';
 
 type Item = string;
