+++ src/components/Radio/__stories__/examples/RadioExampleUse/RadioExampleUse.tsx
@@ -4,7 +4,7 @@ import { action } from '@storybook/addon-actions';
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cnDocsExample } from '../../../../../uiKit/components/DocsExample/DocsExample';
 import { StoryBookExample } from '../../../../../uiKit/components/StoryBookExample/StoryBookExample';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Radio } from '../../../Radio';
 
 const emptyFunction = () => action('emptyFunction');
