+++ src/components/Switch/__stories__/examples/SwitchExampleName/SwitchExampleName.tsx
@@ -4,7 +4,7 @@ import { action } from '@storybook/addon-actions';
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cnDocsExample } from '../../../../../uiKit/components/DocsExample/DocsExample';
 import { StoryBookExample } from '../../../../../uiKit/components/StoryBookExample/StoryBookExample';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Switch } from '../../../Switch';
 
 const emptyFunction = action('emptyFunction');
