+++ src/components/Text/__stories__/Text.stories.tsx
@@ -9,7 +9,7 @@ import {
   textPropDisplay,
   textPropFont,
   textPropFontStyle,
-  textPropHeight,
+  textPropLineHeight,
   textPropSize,
   textPropSizeDefault,
   textPropSpacing,
@@ -30,7 +30,7 @@ const defaultKnobs = () => ({
   decoration: select('decoration', ['', ...textPropDecoration], ''),
   display: select('display', ['', ...textPropDisplay], ''),
   font: select('font', textPropFont, 'primary'),
-  lineHeight: select('lineHeight', ['', ...textPropHeight], ''),
+  lineHeight: select('lineHeight', ['', ...textPropLineHeight], ''),
   size: select('size', textPropSize, textPropSizeDefault),
   spacing: select('spacing', ['', ...textPropSpacing], ''),
   fontStyle: select('fontStyle', ['', ...textPropFontStyle], ''),
