+++ src/components/Text/__stories__/examples/TextExampleDecoration/TextExampleDecoration.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Text } from '../../../Text';
 
 export const TextExampleDecoration = () => {
