+++ src/components/Text/__stories__/examples/TextExampleLineheight/TextExampleLineheight.tsx
@@ -1,18 +1,17 @@
 import React from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
-import { Text, TextPropLineHeight } from '../../../Text';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
+import { Text, textPropLineHeight } from '../../../Text';
 
 export const TextExampleLineheight = () => {
-  const lineHeight: Array<TextPropLineHeight> = ['2xs', 'xs', 's', 'm', 'l'];
   return (
     <div
       className={cnDocsDecorator('Section', [
         wp.tplGrid({ 'xs-columns': 2, 'col-gap': 'full', 'row-gap': 'full' }),
       ])}
     >
-      {lineHeight.map((item, index) => (
+      {textPropLineHeight.map((item, index) => (
         <div key={index} className={wp.tplGrid('fraction', { row: 'third' })}>
           <Text size="s" view="ghost" className={wp.decorator({ 'indent-b': 'xs' })}>
             {`text*line-height*${item}`}
