+++ src/components/Text/__stories__/examples/TextExampleSize/TextExampleSize.tsx
@@ -1,30 +1,17 @@
 import React from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
-import { Text, TextPropSize } from '../../../Text';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
+import { Text, textPropSize } from '../../../Text';
 
 export const TextExampleSize = () => {
-  const size: Array<TextPropSize> = [
-    '2xs',
-    'xs',
-    's',
-    'm',
-    'l',
-    'xl',
-    '2xl',
-    '3xl',
-    '4xl',
-    '5xl',
-    '6xl',
-  ];
   return (
     <div
       className={cnDocsDecorator('Section', [
         wp.tplGrid({ 'xs-columns': 2, 'col-gap': 'full', 'row-gap': 'full' }),
       ])}
     >
-      {size.map((item, index) => (
+      {textPropSize.map((item, index) => (
         <div key={index} className={wp.tplGrid('fraction', { row: 'third' })}>
           <Text size="s" view="ghost" className={wp.decorator({ 'indent-b': 'm' })}>
             {`text*size*${item}`}
