+++ src/components/Text/__stories__/examples/TextExampleSpacing/TextExampleSpacing.tsx
@@ -1,18 +1,17 @@
 import React from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
-import { Text, TextPropSpacing } from '../../../Text';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
+import { Text, textPropSpacing } from '../../../Text';
 
 export const TextExampleSpacing = () => {
-  const spacing: Array<TextPropSpacing> = ['xs', 's', 'm', 'l'];
   return (
     <div
       className={cnDocsDecorator('Section', [
         wp.tplGrid({ 'xs-columns': 2, 'col-gap': 'full', 'row-gap': 'full' }),
       ])}
     >
-      {spacing.map((item, index) => (
+      {textPropSpacing.map((item, index) => (
         <div key={index} className={wp.tplGrid('fraction', { row: 'third' })}>
           <Text size="s" view="ghost" className={wp.decorator({ 'indent-b': 'xs' })}>
             {`text*spacing*${item}`}
