+++ src/components/Text/__stories__/examples/TextExampleView/TextExampleView.tsx
@@ -1,28 +1,17 @@
 import React from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
-import * as wp from '../../../../../utils/whitepaper/whitepaper';
-import { Text, TextPropView } from '../../../Text';
+import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
+import { Text, textPropView } from '../../../Text';
 
 export const TextExampleView = () => {
-  const view: Array<TextPropView> = [
-    'alert',
-    'brand',
-    'ghost',
-    'link',
-    'link-minor',
-    'primary',
-    'secondary',
-    'success',
-    'warning',
-  ];
   return (
     <div
       className={cnDocsDecorator('Section', [
         wp.tplGrid({ 'xs-columns': 3, 'col-gap': 'full', 'row-gap': 'full' }),
       ])}
     >
-      {view.map((item, index) => (
+      {textPropView.map((item, index) => (
         <div key={index} className={wp.tplGrid('fraction', { row: 'third' })}>
           <Text size="s" view="ghost" className={wp.decorator({ 'indent-b': 'xs' })}>
             {`text*view*${item}`}
