--- src/components/Theme/_gap/Theme_gap_m.css
@@ -1,22 +1,0 @@
-.Theme_gap_m {
-  --gap: var(--space-m);
-  --col-gap-third: calc(var(--gap) * 0.33);
-  --col-gap-half: calc(var(--gap) * 0.5);
-  --col-gap-two-thirds: calc(var(--gap) * 0.66);
-  --col-gap-full: var(--gap);
-}
-@media screen and (min-width: 720px) {
-  .Theme_gap_m {
-    --gap: var(--space-2xl);
-  }
-}
-@media screen and (min-width: 1024px) {
-  .Theme_gap_m {
-    --gap: var(--space-3xl);
-  }
-}
-@media screen and (min-width: 1280px) {
-  .Theme_gap_m {
-    --gap: var(--space-4xl);
-  }
-}
