+++ src/components/Theme/presets/presetGpnDisplay.ts
@@ -1,4 +1,3 @@
-import '../../../utils/whitepaper/whitepaper.css';
 import '../Theme.css';
 import '../_color/Theme_color_gpnDefault.css';
 import '../_color/Theme_color_gpnDark.css';
@@ -7,7 +6,6 @@ import '../_control/Theme_control_gpnDefault.css';
 import '../_font/Theme_font_gpnDefault.css';
 import '../_size/Theme_size_gpnDefault.css';
 import '../_space/Theme_space_gpnDefault.css';
-import '../_gap/Theme_gap_m.css';
 
 import { ThemePreset } from '../Theme';
 
@@ -21,5 +19,4 @@ export const presetGpnDisplay: ThemePreset = {
   font: 'gpnDefault',
   size: 'gpnDefault',
   space: 'gpnDefault',
-  gap: 'm',
 };
