+++ src/uiKit/whitepaper/whitepaper.ts
